Changed "Upload folder contains files that were not uploaded" status message on the main form to "Upload folder contains one or more files that were not uploaded".
Changed "UploadStaging folder contains files that were not uploaded" status message on the main form to "UploadStaging folder contains one or more files that were not uploaded".

Changed "Database contains document records that are flagged" status message on the main form to "Database contains one or more document records that are flagged".

Signed-off-by: Robert F. Frasca <rffrasca@gmail.com>